requirements.txt
Django
djangorestframework
asgiref
sqlparse
tzdata
requests
djangorestframework-simplejwt
django-cors-headers
requests
psycopg2-binary
python-dotenv